Moved sprite to new component
