+++ package.json
@@ -18,6 +18,7 @@
     "scss": "^0.2.4"
   },
   "devDependencies": {
+    "@types/node": "^20.10.2",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
     "@types/react-responsive-spritesheet": "^1.0.3",
