+++ src/shared/constants/index.ts
@@ -1,3 +1,10 @@
 export * from "./spriteStates";
 
-export const BASE_SPRITE_SIZE = 128;
+export const ACTOR_STATES = ["Idle", "Walk", "Run", "Hurt", "Dead"];
+
+export const CONFIG = {
+  SPRITE: {
+    BASE_SIZE: 128,
+  },
+  DYNAMIC_STEPS: true,
+};
