+++ src/shared/types/index.ts
@@ -5,8 +5,12 @@ export type StateAttributes = {
 	imgHeight: number;
 };
 
+export type AvailableState = "IDLE" | "WALK" | "RUN" | "HURT" | "DEAD";
+
 export type Sprite = {
 	steps: number;
 	fps: number;
 	goToAndPlay: (param1: number) => void;
 };
+
+export type AvailableActor = "wandererMagician";
